+++ .php-cs-fixer.dist.php
@@ -5,4 +5,5 @@ return Innmind\CodingStandard\CodingStandard::config([
     'tests',
     'fixtures',
     'properties',
+    'proofs',
 ]);