+++ tests/Adapter/FilesystemTest.php
@@ -6,7 +6,6 @@ namespace Tests\Innmind\Filesystem\Adapter;
 use Innmind\Filesystem\{
     Adapter\Filesystem,
     Adapter,
-    CaseSensitivity,
     File,
     File\Content,
     Name,
@@ -438,7 +437,8 @@ class FilesystemTest extends TestCase
                 $name = ".$name";
                 $path = \sys_get_temp_dir().'/innmind/filesystem/';
                 (new FS)->remove($path);
-                (new FS)->dumpFile($path.$name, 'bar');
+                (new FS)->mkdir($path);
+                \file_put_contents($path.$name, 'bar');
 
                 $filesystem = Filesystem::mount(Path::of($path));